feat: add pre-computed waveforms

+++ backend/separation-worker/separation_processor.py
@@ -3,12 +3,14 @@ from prisma import Prisma
 from bullmq import Job
 from minio import Minio
 from dotenv import load_dotenv
-from demucs.api import Separator, save_audio
-import torchaudio as ta
+from demucs.api import Separator
+from demucs.audio import save_audio
+import torchaudio
 import torch as th
 import io
 import uuid
 import shutil
+from utils import generate_waveform_json
 
 # Load environment variables
 load_dotenv()
@@ -39,34 +41,42 @@ class SeparationProcessor:
         await self.prisma.disconnect()
 
     def initialize_minio_client(self):
-        minio_access_key = os.getenv("MINIO_ACCESS_KEY")
-        minio_secret_key = os.getenv("MINIO_SECRET_KEY")
-        minio_endpoint = os.getenv("MINIO_ENDPOINT")
-        minio_port = int(os.getenv("MINIO_PORT", 9000))
-        minio_default_bucket = os.getenv("MINIO_DEFAULT_BUCKET")
+        self.minio_access_key = os.getenv("MINIO_ACCESS_KEY", "")
+        self.minio_secret_key = os.getenv("MINIO_SECRET_KEY", "")
+        self.minio_endpoint = os.getenv("MINIO_ENDPOINT", "")
+        self.minio_port = int(os.getenv("MINIO_PORT", 9000))
+        self.minio_default_bucket = os.getenv("MINIO_DEFAULT_BUCKET", "")
 
         self.minio_client = Minio(
-            f"{minio_endpoint}:{minio_port}",
-            access_key=minio_access_key,
-            secret_key=minio_secret_key,
+            f"{self.minio_endpoint}:{self.minio_port}",
+            access_key=self.minio_access_key,
+            secret_key=self.minio_secret_key,
             secure=False,  # Set to True for HTTPS
         )
 
         # Ensure the default bucket exists
-        print(f"Ensuring bucket {minio_default_bucket} exists...")
-        if not self.minio_client.bucket_exists(minio_default_bucket):
-            self.minio_client.make_bucket(minio_default_bucket)
+        print(f"Ensuring bucket {self.minio_default_bucket} exists...")
+        if not self.minio_client.bucket_exists(self.minio_default_bucket):
+            self.minio_client.make_bucket(self.minio_default_bucket)
 
         print("Minio client initialized.")
 
     async def process(self, job: Job, jobToken):
         await self.connect_to_db()
+        id = None
+        userId = None
+        if not self.minio_client or not self.prisma:
+            raise RuntimeError("Failed to connect to Minio or Prisma")
+
         try:
-            id = job.data["audioFileId"]
-            userId = job.data["userId"]
+            id = str(job.data["audioFileId"])
+            userId = str(job.data["userId"])
             audio_file = await self.prisma.audiofile.find_first(
                 where={"id": id, "userId": userId}
             )
+
+            if not audio_file:
+                raise RuntimeError(f"Audio file with id {id} not found")
 
             print(f"Found audio file: {audio_file.name}, processing...")
 
@@ -74,7 +84,7 @@ class SeparationProcessor:
             file_type = audio_file.fileType
             print(f"File path: {file_path}", f"File type: {file_type}")
             response = self.minio_client.get_object(
-                os.getenv("MINIO_DEFAULT_BUCKET"), file_path
+                self.minio_default_bucket, file_path
             )
 
             try:
@@ -84,17 +94,17 @@ class SeparationProcessor:
                 response.release_conn()
 
             audio_buffer = io.BytesIO(file_data)
-            waveform, sample_rate = ta.load(audio_buffer)
+            waveform, sample_rate = torchaudio.load(audio_buffer)
             print(f"Sample rate: {sample_rate}")
             print("Separating stems...")
             _, separated_stems = self.separator.separate_tensor(waveform)
             print(f"Stems count: {len(separated_stems)}")
 
-            # Save to local file system for now
             # TODO Don't save to local file system, directly save to object storage from stream
             for stem, source in separated_stems.items():
                 stem_name = uuid.uuid4()
-                full_path = f"{userId}/{stem_name}.{file_type}"
+                full_name = f"{userId}/{stem_name}"
+                full_path = f"{full_name}.{file_type}"
                 # check if userId folder exists
                 if not os.path.exists(userId):
                     os.makedirs(userId)
@@ -104,20 +114,23 @@ class SeparationProcessor:
 
                 # Save to minio
                 self.minio_client.fput_object(
-                    os.getenv("MINIO_DEFAULT_BUCKET"),
+                    os.getenv("MINIO_DEFAULT_BUCKET") or "audio",
                     full_path,
                     file_path=full_path,
                     metadata={"mimeType": f"audio/{file_type}", "name": stem},
                 )
 
+                waveform_json = await generate_waveform_json(full_path)
+
                 # Save to database
                 response = await self.prisma.audiofile.create(
-                    {
+                    data={
                         "name": stem,
                         "filePath": full_path,
-                        "fileType": "wav",
                         "userId": userId,
                         "parentId": id,
+                        "fileType": file_type,
+                        "waveform": waveform_json,
                     }
                 )
 
@@ -135,7 +148,8 @@ class SeparationProcessor:
         except Exception as e:
             print(e)
             await self.disconnect_from_db()
-            shutil.rmtree(userId)
+            if userId and os.path.exists(userId):
+                shutil.rmtree(userId)
             result = {
                 "userId": userId,
                 "audioFileId": id,
